Move classes to utils subpackage
